Update README.md
EDA on General Elections Dataset

Samples were collected over a period of nine days where the respondents were asked questions related to results of general elections. Survey weights are in the 'weight' column. The data analysis will be done with the help of some questions with the help of few libraries and visualization is done accordingly.